Refactor code structure.
