+++ Cargo.toml
@@ -7,6 +7,10 @@ homepage = "https://www.omnity.network/"
 repository = "https://github.com/octopus-network/ree-types"
 description = "Data type definitions of REE."
 
+[lib]
+name = "ree_types"
+crate-type = ["lib"]
+
 [dependencies]
 candid = "0.10"
 ic-cdk = "0.17"
@@ -16,5 +20,5 @@ generic-array = "0.14"
 ic-stable-structures = "0.6"
 serde = { version = "1.0", features = ["derive"] }
 hex = "0.4"
-bitcoin = { version = "0.32.4", default-features = false, features = ["base64", "serde"] }
-ciborium = "0.2.1"
+bitcoin = { version = "0.32", default-features = false, features = ["base64", "serde"] }
+ciborium = "0.2"
